Refactor Dialogs to Table vs det_msg
Implements a custom dialog for the sync logs including a tabular log view for readability.

Note: While functional I think a conversation around what is actually relevant in the user log would be beneficial.

Also implements silent as a bool param for the 3 sync function to not generate a user dialog. Updated automatic device sync and scheduled ProgressSync to be silent

Does not include icon/image for the dialog, can be added in if needed/wanted (info/warn/error/etc)