Implement show_restart_warning in config
The settings that require a restart are checked for changes at save and prompt user for restart with a restart button.

Had to split out the automatic device sync from CustomCheckboxLayout to be able to read the value easily. The formatting is off (checkbox left/right, tooltip, etc.) so ideally it goes back into CustomCheckboxLayout, might need some help with that.